requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gcloud==0.17.0
google-api-core==1.26.1
google-auth==1.27.1
google-cloud-vision==2.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jws==0.1.3
libcst==0.3.17
MarkupSafe==1.1.1
mypy-extensions==0.4.3
oauth2client==3.0.0
packaging==20.9
proto-plus==1.17.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug==1.0.1
